Rename the DragActivityTool to GameTool
