+++ src/BloomBrowserUI/bookEdit/toolbox/dragActivity/GameTool.tsx
